Text position fixed

+++ SiroGen/scene.cpp
@@ -37,7 +37,7 @@ void Scene::DeleteEntity(Entity* entity)
 	std::cout << "Warning: DeleteEntity is unfinished" << std::endl;
 }
 
-void Scene::AddTexttoScene(const char* text, int x, int y, int size, const char* TGAfont)
+void Scene::AddTexttoScene(const char* text, float x, float y, float size, const char* TGAfont)
 {
 	Text* temptext = new Text();
 	temptext->printText2D(text, x, y, size, TGAfont);
