+++ SiroGen/text.h
@@ -8,7 +8,7 @@ class Text : public Component
 {
 public:
 	void initText2D(const char* texturePath);
-	void printText2D(const char* text, int x, int y, int size, const char* TGAfont = "assets/LucidaConsole.tga");
+	void printText2D(const char* text, float x, float y, float size, const char* TGAfont = "assets/LucidaConsole.tga");
 	void cleanupText2D();
 	virtual void DoIt(GLuint shader);
